requirements.txt
python-telegram-bot==13.7
telethon==1.34.0
firebase-admin==6.5.0
beautifulsoup4==4.12.3
aiohttp==3.9.3
python-dotenv==1.0.1
lxml==5.1.0
google-generativeai==0.8.3
brotli
requests
curl_cffi==0.7.4
